Rq 클래스 도입

+++ src/main/java/com/domain/wiseSaying/controller/WiseSayingController.java
@@ -3,11 +3,9 @@ package com.domain.wiseSaying.controller;
 import com.AppContext;
 import com.domain.wiseSaying.entity.WiseSaying;
 import com.domain.wiseSaying.service.WiseSayingService;
+import com.global.Rq;
 
-import java.util.Arrays;
-import java.util.Map;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 public class WiseSayingController {
     private final Scanner scanner;
@@ -43,19 +41,13 @@ public class WiseSayingController {
         }
     }
 
-    public void actionDelete(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
+    public void actionDelete(Rq rq) {
+        int id = rq.getParamAsInt("id", -1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e-> e.split("=", 2))
-                .filter(e-> e.length == 2 && !e[0].isEmpty() && !e[1].isEmpty())
-                .collect(Collectors.toMap(e-> e[0].trim(), e -> e[1].trim()));
-
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
-
+        if(id == -1){
+            System.out.println("id를 정확히 입력해주세요");
+            return;
+        }
         boolean deleted = wiseSayingService.delete(id);
         if(!deleted){
             System.out.printf("%d번 명언은 존재하지 않습니다.\n", id);
@@ -64,18 +56,8 @@ public class WiseSayingController {
         System.out.printf("%d번 명언이 삭제되었습니다.\n", id);
     }
 
-    public void actionModify(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
-
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e -> e.split("=", 2))
-                .filter(e -> e.length == 2 && !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e -> e[0].trim(), e -> e[1].trim()));
-
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
+    public void actionModify(Rq rq) {
+        int id = rq.getParamAsInt("id", -1);
 
         WiseSaying wiseSaying = wiseSayingService.findById(id);
